Added a way to prevent negative value and float values

+++ inc/Engine/Preload/Database/Queries/Cache.php
@@ -277,11 +277,11 @@ class Cache extends Query {
 	/**
 	 * Get all preload caches which were not accessed in the last month.
 	 *
-	 * @param int    $delay delay before the not accessed row is deleted.
+	 * @param float  $delay delay before the not accessed row is deleted.
 	 * @param string $unit unit from the delay.
 	 * @return array
 	 */
-	public function get_old_cache( int $delay = 1, string $unit = 'month' ) : array {
+	public function get_old_cache( float $delay = 1, string $unit = 'month' ) : array {
 		// Get the database interface.
 		$db = $this->get_db();
 
@@ -300,11 +300,11 @@ class Cache extends Query {
 	/**
 	 * Remove all completed rows one by one.
 	 *
-	 * @param int    $delay delay before the not accessed row is deleted.
+	 * @param float  $delay delay before the not accessed row is deleted.
 	 * @param string $unit unit from the delay.
 	 * @return void
 	 */
-	public function remove_all_not_accessed_rows( int $delay = 1, string $unit = 'month' ) {
+	public function remove_all_not_accessed_rows( float $delay = 1, string $unit = 'month' ) {
 		$rows = $this->get_old_cache( $delay, $unit );
 
 		foreach ( $rows as $row ) {
